Extracted seat layout to SeatSelector component with dynamic salong and wheelchair seats
